Rename the original utilities.js to common.js to indicate its generic use
