+++ settings/src/components/backup-codes.js
@@ -10,7 +10,7 @@ import { Icon, warning, cancelCircleFilled } from '@wordpress/icons';
  * Internal dependencies
  */
 import { GlobalContext } from '../script';
-import { refreshRecord } from '../utilities';
+import { refreshRecord } from '../utilities/common';
 
 /**
  * Setup and manage backup codes.
