+++ settings/src/components/webauthn/list-keys.js
@@ -9,8 +9,7 @@ import { Icon, cancelCircleFilled } from '@wordpress/icons';
  * Internal dependencies
  */
 import { GlobalContext } from '../../script';
-import { refreshRecord } from '../../utilities';
-
+import { refreshRecord } from '../../utilities/common';
 /**
  * Render the list of keys.
  */
