+++ settings/src/utilities/common.js
